feat(web): render work activity feed as a table matching /activity
Replace the FeedRow card list on /works/[id]/activity with a FeedTable
that mirrors the global activity-log table design (bordered container,
uppercase header, Date/Time | Type | Summary | Status columns, shared
status badge and stacked timestamp). Extract ActivityTimestamp into a
shared component, restyle the loading skeleton to the table shape, and
port the FeedRow specs (incl. the safeExternalUrl XSS guard) to the new
component.
